Add custom parsers for command context values
