+++ Core/src/main/java/com/darwinreforged/server/core/commands/context/parsing/AbstractParser.java
@@ -1,0 +1,3 @@
+package com.darwinreforged.server.core.commands.context.parsing;
+
+public interface AbstractParser {}
