+++ Core/src/main/java/com/darwinreforged/server/core/commands/context/parsing/ArgumentParser.java
@@ -1,0 +1,11 @@
+package com.darwinreforged.server.core.commands.context.parsing;
+
+import com.darwinreforged.server.core.commands.context.AbstractCommandValue;
+
+import java.util.Optional;
+
+public abstract class ArgumentParser implements AbstractParser {
+
+    public abstract <A> Optional<A> parse(AbstractCommandValue<String> commandValue);
+
+}
